requirements.txt
pyspark==3.4.0
pymongo==4.3.3
kafka-python==2.0.2
Flask==2.2.2
numpy==1.24.1
pandas==1.5.3
virtualenv==20.21.0
